requirements.txt
tensorflow==2.14.0
numpy
pandas
matplotlib
dvc
dvc-s3
dvclive
awscli
jupyterlab
python-dotenv
psycopg2-binary